Slight refactor / naming fixes. Reverted unneeded change to UpdateCorrespondenceStatusHelper.ValidateUpdateRequest
