+++ src/Altinn.Correspondence.Application/SyncCorrespondenceEvent/SyncCorrespondenceStatusRequest.cs
@@ -1,7 +1,7 @@
 using Altinn.Correspondence.Core.Models.Entities;
 using Altinn.Correspondence.Core.Models.Enums;
 
-namespace Altinn.Correspondence.Application.SyncCorrespondenceStatusEvent;
+namespace Altinn.Correspondence.Application.SyncCorrespondenceEvent;
 
 public class SyncCorrespondenceStatusEventRequest
 {
